LayoutTests/accessibility/empty-image-with-title.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<img id="image1" src="foo.gif" title="baz" width="100" height="100">

<div id="console"></div>

<script>

    description("This tests that if a missing image has a title attribute, it still appears in the AX hierarchy.");

    if (window.accessibilityController) {

          var image1 = accessibilityController.accessibleElementById("image1");
          shouldBe("image1.role", "'AXRole: AXImage'");
          shouldBe("image1.description", "'AXDescription: baz'");
    }

</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
